# www_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the Laravel application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pet-types": {
            "get": {
                "tags": [
                    "PetTypes"
                ],
                "summary": "Get list of Pet Types",
                "operationId": "f7c75ad6326bb9c3251720f9243e618b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "собака"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-12T20:08:04.566Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "PetTypes"
                ],
                "summary": "Store a new Pet Type",
                "description": "Creates a new Pet Type and stores it in the database.",
                "operationId": "storePetTypes",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_uk",
                                    "name_en",
                                    "name_ru"
                                ],
                                "properties": {
                                    "name_uk": {
                                        "description": "Name of the created Pet Type in Ukrainian",
                                        "type": "string",
                                        "example": "собака"
                                    },
                                    "name_en": {
                                        "description": "Name of the created Pet Type in English",
                                        "type": "string",
                                        "example": "dog"
                                    },
                                    "name_ru": {
                                        "description": "Name of the created Pet Type in Russian",
                                        "type": "string",
                                        "example": "собака"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pet Type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the created Pet Type",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name_uk": {
                                            "description": "Name of the created Pet Type in Ukrainian",
                                            "type": "string",
                                            "example": "собака"
                                        },
                                        "name_en": {
                                            "description": "Name of the created Pet Type in English",
                                            "type": "string",
                                            "example": "dog"
                                        },
                                        "name_ru": {
                                            "description": "Name of the created Pet Type in Russian",
                                            "type": "string",
                                            "example": "собака"
                                        },
                                        "updated_at": {
                                            "description": "Date and time of the last update",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:08:04.566Z"
                                        },
                                        "created_at": {
                                            "description": "Date and time of the creation",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:08:04.566Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}